Refactor: reorganizing assets
