+++ warehouse/oso_dagster/cbt/utils/replace.py
@@ -1,3 +1,5 @@
+from typing import cast
+
 from sqlglot import expressions as exp
 from .compare import is_same_source_table
 
@@ -6,6 +8,7 @@ def replace_source_tables(search: exp.Table, replace: exp.Table):
     def _transform(expression: exp.Expression):
         if type(expression) not in [exp.Table]:
             return expression
+        expression = cast(exp.Table, expression)
         if not is_same_source_table(search, expression):
             return expression
         replacement = replace.copy()
